Added basic interfaces for observer design pattern
