+++ main-principles.txt
@@ -3,3 +3,4 @@ Main design principles:
 1. Identify the aspects of your application that vary and separate them from what stays the same.
 2. Program to an interface, not an implementation.
 3. Favor composition over inheritance.
+4. Strive for loosely coupled designs between objects that interact.
