+++ src/main/java/design/object/observer/Display.java
@@ -1,0 +1,12 @@
+package design.object.observer;
+
+/**
+ * Defines how an object looks
+ */
+public interface Display {
+
+    /**
+     * Displays an object
+     */
+    void display();
+}
