+++ src/main/java/design/object/strategy/behavior/FlyBehavior.java
@@ -1,4 +1,4 @@
-package design.object.chapter01.behavior;
+package design.object.strategy.behavior;
 
 /**
  * Interface for flying capabilities
