+++ src/main/java/design/object/strategy/behavior/FlyNoWay.java
@@ -1,4 +1,4 @@
-package design.object.chapter01.behavior;
+package design.object.strategy.behavior;
 
 /**
  * For creatures who aren't able to fly
