+++ src/main/java/design/object/strategy/behavior/QuackBehavior.java
@@ -1,4 +1,4 @@
-package design.object.chapter01.behavior;
+package design.object.strategy.behavior;
 
 /**
  * Interface for quacking capabilities
